Fixed code
Fixed constructors:
You shouldnt delete pointers in constructors, it is not possible to call
constructors again after the object is already created.
So removed delete operations -> they will crash when they try to delete
a pointer that doesn't exist.

Used char pointer:
Created an integer variable to track the len of the char pointer/array.